etl/requirements.txt
pyspark
pandas
python-dotenv